oct-dataset/requirements.txt
matplotlib>=3.8.4
numpy>=1.26.4
opencv_python>=4.9.0.80
pandas>=2.2.2
Pillow>=10.3.0
scipy>=1.13.1
torch>=2.3.0
torchvision>=0.18.0
tqdm>=4.66.2
ultralytics>=8.2.31
